refactor: move file and refactor structure
