+++ app/src/main/java/com/weiran/mynowinandroid/MainActivity.kt
@@ -3,7 +3,7 @@ package com.weiran.mynowinandroid
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.weiran.mynowinandroid.ui.page.Home
+import com.weiran.mynowinandroid.home.ui.Home
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
