+++ app/src/main/java/com/weiran/mynowinandroid/theme/MyIcons.kt
@@ -1,11 +1,10 @@
-package com.weiran.mynowinandroid.ui.theme
+package com.weiran.mynowinandroid.theme
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Add
 import androidx.compose.material.icons.rounded.Check
 import androidx.compose.material.icons.rounded.Settings
 import androidx.compose.material.icons.rounded.Share
-import com.weiran.mynowinandroid.R
 
 object MyIcons {
     val Add = Icons.Rounded.Add
